+++ changelog.md
@@ -3,7 +3,7 @@
 *Everything unchecked is currently not supported in the front end*
 
 - [ ] batch api
-- [ ] getPlayer command extends by looking_at and sneak
+- [X] getPlayer command extends by looking_at and sneak
 - [ ] editEntity command
 - [ ] setPlayerStat command
 - [ ] setPlayerVelocity command